fix: count token
